refactor: consolidate favorites logic and standardize file structure
- Move 'api.ts' to 'services/' and rename 'setupTests.ts' to kebab-case for consistency.
- Centralize 'changeFavoriteStatusAction' in favorites thunks to adhere to DRY principle.
- Remove duplicate logic from offer thunks.
- Update 'use-favorites' hook to match the new action signature.
