+++ src/store/offer/offer-thunks.ts
@@ -79,27 +79,3 @@ export const postCommentAction = createAsyncThunk<
     return rejectWithValue(axiosError.response?.data?.message || 'Failed to post comment');
   }
 });
-
-export const changeFavoriteStatusAction = createAsyncThunk<
-  Offer,
-  { offerId: string; status: number },
-  {
-    dispatch: AppDispatch;
-    state: State;
-    extra: AxiosInstance;
-    rejectValue: string;
-  }
->('offer/changeFavoriteStatus', async ({ offerId, status }, { extra: api, rejectWithValue }) => {
-  try {
-    const { data } = await api.post<Offer>(`${APIRoute.Favorite}/${offerId}/${status}`);
-    return data;
-  } catch (error) {
-    const axiosError = error as AxiosError<CommonError>;
-
-    if (axiosError.response?.status === StatusCodes.UNAUTHORIZED) {
-      return rejectWithValue('UNAUTHORIZED');
-    }
-
-    return rejectWithValue(axiosError.response?.data?.message || 'Failed to change favorite status');
-  }
-});
